Titter refactoring, ES6 introduction, prepare for webgold donate integration
