+++ .travis.yml
@@ -1,7 +1,7 @@
 language: node_js
 
 node_js:
-  - 0.10
+  - 0.12
 
 before_install:
   - sudo apt-get install libcairo2-dev libjpeg8-dev libpango1.0-dev libgif-dev build-essential g++
