+++ package.json
@@ -17,6 +17,7 @@
     "ejs": "^1.0.0",
     "htm": "*",
     "es6-promise": "^2.1.1",
+    "es6-promisify":"*",
     "express": "^4.10.1",
     "express-session": "*",
     "gm": "^1.17.0",
@@ -32,7 +33,9 @@
     "mocha": "^2.2.5",
     "multer": "*",
     "should": "^6.0.3",
-    "supertest": "^1.0.1"
+    "supertest": "^1.0.1",
+    "babel": "^5.6.14",
+    "babelify": "^6.1.2"
   },
   "scripts": {
     "jsx": "jsx --no-cache-dir -x jsx ./widget ./js",
@@ -40,7 +43,8 @@
     "watch": "jsx --no-cache-dir -w -x jsx ./widget ./js",
     "bumpVersion": "npm run version && npm run bundle",
     "version": "sh version.sh",
-    "bundle": "sh bundle.sh"
+    "bundle": "sh bundle.sh",
+    "test": "export NODE_ENV=testing && mocha --timeout 20000 --compilers js:babel/register"
   },
   "repository": {
     "type": "git",
